# swagger.yaml
openapi: 3.0.3
info:
  title: Education Platform API
  description: >-
    API for managing students, teachers, categories, courses, articles,
    discount codes, subscriptions and user dashboard **with OTP‑based
    authentication** (no password).  
    **Version 2.0.0** – completely password‑less flow.
  version: 2.0.0
servers:
  - url: http://localhost:5000/api
    description: Local server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    Error:
      type: object
      properties:
        message: { type: string }
      required: [message]
...
    Comment:
      type: object
      properties:
        user: { type: string }
        text: { type: string }
        rating: { type: number, minimum: 0, maximum: 5 }
        status: { type: string, enum: [pending, approved], default: pending }
        createdAt: { type: string, format: date-time }
      required: [user, text, status]
    Course:
      type: object
      properties:
        _id: { type: string }
        coverImage: { type: string }
        description: 
          type: string
          example: "در این دوره با مفاهیم پایه تا پیشرفته React آشنا می‌شوید. شامل پروژه‌های واقعی، Hooks، Redux و..."
        title: { type: string }
        category: { type: string }
        teacher: { type: string }
        status: { type: string, enum: [active, pending, pre-register] }
        level: { type: string, enum: [beginner, intermediate, advanced] }
        duration: { type: number }
        previewVideo: { type: string }
        presentationMethod: { type: string, enum: [download, streaming] }
        downloadLink: { type: string }
        prerequisites: { type: array, items: { type: string } }
        type: { type: string, enum: [free, vip, paid] }
        price: { type: number }
        discount: { type: number }
        chapters:
          type: array
          items:
            type: object
            properties:
              title: { type: string }
              duration: { type: number }
              description: { type: string }
              videos:
                type: array
                items:
                  type: object
                  properties:
                    title: { type: string }
                    description: { type: string }
                    duration: { type: number }
                    time: { type: string }
                    streamingLink: { type: string }
                  required: [title, duration]
            required: [title, duration]
        students: { type: array, items: { type: string } }
        rating: { type: number }
        expertise: { type: string }
        comments:
          type: array
          items: { $ref: '#/components/schemas/Comment' }
      required:
        - coverImage
        - title
        - category
        - teacher
        - status
        - level
        - duration
        - presentationMethod
        - type
    #-----------------------
...
    PaymentHistoryItem:
      type: object
      properties:
        paymentId:
          type: string
          description: شناسه پرداخت
        amount:
          type: number
          description: مبلغ کل
        status:
          type: string
          enum: [pending, completed, failed]
        refId:
          type: string
          nullable: true
          description: شماره پیگیری بانک
        createdAt:
          type: string
          format: date-time
        itemsCount:
          type: integer
          description: تعداد آیتم‌ها (دوره + اشتراک)
        items:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
                enum: [course, subscription]
              title:
                type: string
              coverImage:
                type: string
                format: uri
                nullable: true
              id:
                type: string
            required: [type, title, id]
        user:
          type: object
          nullable: true
          description: فقط برای ادمین
          properties:
            id: { type: string }
            name: { type: string }
            phone: { type: string }
            email: { type: string }
      required: [paymentId, amount, status, createdAt, itemsCount, items]

    # -------------------------------------------------
    # USER DASHBOARD (به‌روز شده)
    # -------------------------------------------------
...
  responses:
      Unauthorized:
        description: Invalid or missing token
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Error' }
      Forbidden:
        description: Admin access required
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Error' }
...
      NotFound:
        description: Resource not found
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Error' }
      ServerError:
        description: Internal server error
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Error' }

# ==============================================================
# ===========================  PATHS  ==========================
# ==============================================================
paths:
...
  /payments:
    post:
      summary: Create a payment for a course
      tags: [Payments]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                courseId: { type: string }
                discountCode: { type: string }
              required: [courseId]
      responses:
        '200':
          description: Payment created or enrolled without payment
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  paymentUrl: { type: string }
                  authority: { type: string }
                  paymentId: { type: string }
                  originalPrice: { type: number }
                  courseDiscountPercent: { type: number }
                  courseDiscountAmount: { type: number }
                  codeDiscountPercent: { type: number }
                  codeDiscountAmount: { type: number }
                  finalPrice: { type: number }
                  paymentRequired: { type: boolean }
        '400':
          description: Bad request
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: Course or user not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '500':
          description: Server error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /payments/verify:
    get:
      summary: Verify a payment
      tags: [Payments]
      parameters:
        - in: query
          name: paymentId
          schema: { type: string }
          required: true
        - in: query
          name: refId
          schema: { type: string }
          required: true
        - in: query
          name: status
          schema: { type: string }
          required: true
      responses:
        '200':
          description: Payment verified
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  refId: { type: string }
                  courseId: { type: string }
                  subscriptionPlanId: { type: string }
                  subscriptionExpiresAt: { type: string, format: date-time }
        '400':
          description: Payment failed or already processed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: Payment not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '500':
          description: Server error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  #--------------------------
  /payments/history:
    get:
      summary: Get payment history

      tags: [Payments]
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: userId
          schema: { type: string }
          description: فقط برای ادمین – شناسه کاربر
      responses:
        '200':
          description: لیست تاریخچه پرداخت
          content:
            application/json:
              schema:
                type: object
                properties:
                  payments:
                    type: array
                    items: { $ref: '#/components/schemas/PaymentHistoryItem' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/ServerError' }
  /payments/history/user/{userId}:
    get:
      summary: Get payment history of a specific user (Admin only)
      tags: [Payments, Admin]
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      responses:
        '200':
          description: پرداخت‌های کاربر خاص
          content:
            application/json:
              schema:
                type: object
                properties:
                  payments:
                    type: array
                    items: { $ref: '#/components/schemas/PaymentHistoryItem' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/ServerError' }
  /payments/history/all:
    get:
      summary: Get all payment history (Admin only)
      tags: [Payments, Admin]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: همه پرداخت‌ها
          content:
            application/json:
              schema:
                type: object
                properties:
                  payments:
                    type: array
                    items: { $ref: '#/components/schemas/PaymentHistoryItem' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }


  # -------------------  HOMEPAGE  -------------------
  /homepage:
    get:
      summary: Get homepage data with statistics and paginated courses
      tags: [Homepage]

      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  statistics:
                    type: object
                    properties:
                      totalCourses: { type: integer }
                      completedCourses: { type: integer }
                      ongoingCourses: { type: integer }
                      totalStudents: { type: integer }
                      latestCourseTime: { type: string }
                  courses:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string }
                        title: { type: string }
                        coverImage: { type: string }
                        teacher: { type: string }
                        expertise: { type: string }
                        price: { type: number }
                        status: { type: string }
                        level: { type: string }
                        duration: { type: number }
                        type: { type: string }
                        category: { type: string }
                        createdAt: { type: string, format: date-time }
                  pagination:
                    type: object
                    properties:
                      currentPage: { type: integer }
                      totalPages: { type: integer }
                      totalCourses: { type: integer }
                      limit: { type: integer }
                  canCreateCourse: { type: boolean }
        '500':
          description: Server error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /homepage/search:
    get:
      summary: Course autocomplete / suggestions
      description: |
        Returns up to 5 active courses that match the query on **title**, **teacher name**, or **category name**.
        Each result contains **full course data** (same fields as `GET /courses/{id}`) so the front-end can navigate directly to the course detail page without an extra request.
      tags: [Homepage]
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: q
          schema:
            type: string
            minLength: 2
          required: true
          description: Search term (minimum 2 characters)
          example: react
      responses:
        '200':
          description: List of matching courses (full detail)
          content:
            application/json:
              schema:
                type: object
                properties:
                  suggestions:
                    type: array
                    maxItems: 5
                    items:
                      allOf:
                        - $ref: '#/components/schemas/Course'
                        - type: object
                          properties:
                            teacher:
                              type: string
                              description: Full teacher name (first + last)
                              example: Ali Mohammadi
                            category:
                              type: string
                              description: Category name
                              example: Front-end
                            finalPrice:
                              type: number
                              description: Price after applying course discount
                              example: 280000
                            coverImage:
                              type: string
                              format: uri
                              description: Full URL of the cover image
                            previewVideo:
                              type: string
                              format: uri
                              nullable: true
                              description: Full URL of the preview video
                            chapters:
                              type: array
                              items:
                                type: object
                                properties:
                                  title: { type: string }
                                  duration: { type: number }
                                  description: { type: string }
                                  videos:
                                    type: array
                                    items:
                                      type: object
                                      properties:
                                        title: { type: string }
                                        duration: { type: number }
                                        fileUrl:
                                          type: string
                                          format: uri
                                          nullable: true
                          required:
                            - _id
                            - title
                            - teacher
                            - category
                            - price
                            - discount
                            - finalPrice
                            - coverImage
        '400':
          description: Query string is missing or too short
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'


  # -------------------  BASKET  -------------------
...
security:
  - BearerAuth: []